add una sola card para todo
