+++ src/main.jsx
@@ -4,8 +4,9 @@ import Pokemon from "./pages/Pokemon";
 import Home from "./pages/Home";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import PokemonInicial from "./pages/PokeInicial";
-import PokemonGatcha from "./pages/Gatcha";
-import PokeCardCompleta from "./components/CardCompleta";
+import PokeDescripcion from "./pages/PokeGatcha";
+import PokeGatcha from "./pages/PokeGatcha";
+import PokeCardStat from "./pages/PokeDescripcion";
 // <React.StrictMode>
 //<App />
 // </React.StrictMode>,
@@ -21,21 +22,25 @@ const routes = createBrowserRouter([
       },
       {
         path: "/pokemonGatcha",
-        element: <PokemonGatcha />,
+        element: <PokeGatcha />,
       },
       {
         path: "/pokemonInicial",
         element: <PokemonInicial />,
       },
       {
-        path: "/pokemonCardCompleta/:name",
-        element: <PokeCardCompleta />,
+        path: "/pokemonCardStat/:name",
+        element: <PokeCardStat />,
       },
     ],
   },
   {
     path: "/login",
     element: <p>hola♥</p>,
+  },
+  {
+    path: "/descrip",
+    element: <PokeDescripcion />,
   },
 ]);
