+++ src/pages/PokeGatcha.jsx
@@ -1,17 +1,16 @@
 import { useEffect, useState } from "react";
 import PokeCardCompletaPrueba from "../components/CardCompletaPrueba";
 
-export default function PokemonInicial() {
+export default function PokeGatcha() {
   const [pokemon, setPokemon] = useState({});
 
-  const PokeRandom = () => {
-    const poke = ["1", "4", "7"];
-    const pokeRandom = Math.floor(Math.random() * poke.length);
-    return poke[pokeRandom];
+  const numRandom = (min, max) => {
+    var random = Math.floor(Math.random() * (max - min) + min);
+    return random;
   };
 
   useEffect(() => {
-    fetch(`https://pokeapi.co/api/v2/pokemon/${PokeRandom()}`)
+    fetch(`https://pokeapi.co/api/v2/pokemon/${numRandom(1, 185)}`)
       .then((response) => response.json())
       .then((json) => {
         setPokemon(json);
